api: Rename api.go -> server.go and requests.go -> api.go
